Abstract `StakeVersionContainer`
