+++ src/hooks/useOldAssetsEnoughToMigrate.ts
@@ -1,0 +1,13 @@
+import { useAppSelector } from ".";
+
+export const useOldAssetsEnoughToMigrate = () => {
+  return useAppSelector(state => {
+    if (!state.app.currentIndex || !state.app.marketPrice) {
+      return true;
+    }
+    const wrappedBalance = Number(state.account.balances.wsohm) * Number(state.app.currentIndex!);
+    const allAssetsBalance =
+      Number(state.account.balances.sohmV1) + Number(state.account.balances.ohmV1) + wrappedBalance;
+    return state.app.marketPrice * allAssetsBalance >= 10;
+  });
+};
